Function.py keyword refactor
