+++ p.k
@@ -1,4 +1,4 @@
-//#define __CBLIB_SOURCE
+#define __CBLIB_SOURCE
 #include "cblib.cb"
 
 template<typename T>
@@ -12,7 +12,7 @@ struct Test{
 int main(int argc){
     
     Test<double> test;
-
+    printf("test123: %s%i\n", "test", 123);
 
     return 0;
 }
